Fixed sample count error & refactored
